Move debug into test code
