+++ tests/common/debug.py
