+++ tests/datetime/test_dates.py
@@ -17,9 +17,9 @@
 
 from __future__ import division, print_function, absolute_import
 
+from tests.common.debug import minimal
 from hypothesis.strategytests import strategy_test_suite
 from hypothesis.extra.datetime import dates
-from hypothesis.internal.debug import minimal
 from hypothesis.internal.compat import hrange
 
 TestStandardDescriptorFeatures1 = strategy_test_suite(dates())
